llm_api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
blobfile==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.8
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
filelock==3.12.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.106
llama-index==0.4.24
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.0
openai==0.27.8
packaging==23.1
pandas==2.0.2
psycopg2-binary==2.9.6
pycryptodomex==3.18.0
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.48
sqlparse==0.4.4
tenacity==8.2.2
tiktoken==0.3.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
